# feat: add assembly based interpreter loop to speeeup VM executions