--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,5 @@
 /target
+/definitions/target
 **/*.rs.bk
 Cargo.lock
 .deps/