--- a/build.rs
+++ b/build.rs
@@ -1,10 +1,16 @@
-#[cfg(feature = "jit")]
+#[cfg(any(feature = "asm", feature = "jit"))]
 use cc::Build;
 
 fn main() {
+    #[cfg(any(feature = "asm", feature = "jit"))]
+    let mut build = Build::new();
+
+    #[cfg(all(unix, target_pointer_width = "64", feature = "asm"))]
+    build.file("src/machine/asm/execute.S");
+
     #[cfg(all(unix, target_pointer_width = "64", feature = "jit"))]
-    Build::new()
-        .file("src/jit/asm.x64.compiled.c")
-        .include("dynasm")
-        .compile("asm");
+    build.file("src/jit/asm.x64.compiled.c").include("dynasm");
+
+    #[cfg(any(feature = "asm", feature = "jit"))]
+    build.compile("asm");
 }