--- a/definitions/src/asm.rs
+++ b/definitions/src/asm.rs
@@ -10,6 +10,7 @@ pub const RET_DECODE_TRACE: u8 = 1;
 pub const RET_ECALL: u8 = 2;
 pub const RET_EBREAK: u8 = 3;
 pub const RET_MAX_CYCLES_EXCEEDED: u8 = 4;
+pub const RET_OUT_OF_BOUND: u8 = 5;
 
 #[inline(always)]
 pub fn calculate_slot(addr: u64) -> usize {