--- a/definitions/src/generate_asm_constants.rs
+++ b/definitions/src/generate_asm_constants.rs
@@ -1,10 +1,11 @@
 use ckb_vm_definitions::{
     asm::{
-        AsmCoreMachine, Trace, RET_DECODE_TRACE, RET_EBREAK, RET_ECALL, RET_MAX_CYCLES_EXCEEDED,
+        AsmCoreMachine, Trace, RET_DECODE_TRACE, RET_EBREAK, RET_ECALL, RET_MAX_CYCLES_EXCEEDED, RET_OUT_OF_BOUND,
         TRACE_ITEM_LENGTH,
     },
     instructions::{Instruction, INSTRUCTION_OPCODE_NAMES},
     registers::SP,
+    RISCV_MAX_MEMORY,
 };
 use std::mem::{size_of, zeroed};
 
@@ -16,6 +17,12 @@ use std::mem::{size_of, zeroed};
 // of this as a workaround to the problem that build.rs cannot depend on any
 // of its crate contents.
 fn main() {
+    println!(
+        "#define CKB_VM_ASM_RISCV_MAX_MEMORY {}",
+        RISCV_MAX_MEMORY,
+    );
+    println!();
+
     println!(
         "#define CKB_VM_ASM_MAXIMUM_TRACE_ADDRESS_LENGTH {}",
         TRACE_ITEM_LENGTH * 4
@@ -29,6 +36,7 @@ fn main() {
         "#define CKB_VM_ASM_RET_MAX_CYCLES_EXCEEDED {}",
         RET_MAX_CYCLES_EXCEEDED
     );
+    println!("#define CKB_VM_ASM_RET_OUT_OF_BOUND {}", RET_OUT_OF_BOUND);
     println!();
 
     println!("#define CKB_VM_ASM_REGISTER_SP {}", SP);