--- a/src/jit/instructions.rs
+++ b/src/jit/instructions.rs
@@ -1,4 +1,5 @@
-use crate::instructions::{extract_opcode, Instruction, OP_EBREAK, OP_ECALL};
+use crate::instructions::{extract_opcode, Instruction};
+use ckb_vm_definitions::instructions::{OP_EBREAK, OP_ECALL};
 
 pub fn is_jitable_instruction(i: Instruction) -> bool {
     match extract_opcode(i) {