… .travis.yml
… Cargo.toml
… definitions/Cargo.toml
… src/instructions/mod.rs
… src/lib.rs
… src/machine/mod.rs
… src/machine/trace.rs
… src/memory/mmu.rs
… tests/programs/write_large_address.S
… tests/programs/write_large_address64
… tests/test_asm.rs
… tests/test_jit.rs
… tests/test_misc.rs
… tests/test_simple.rs