This PR adds a new assembly based interpreter loop to speed the VM in x86_64 architectures. While running our [secp256k1-sha3 benchmark](https://github.com/nervosnetwork/ckb-vm-bench-scripts/blob/master/c/secp256k1_bench.c)

* Pure Rust based interpreter: ~60ms
* Rust based interpreter with trace optimization: ~27ms
* Baseline JIT: ~9ms
* This PR: ~7.9ms

Note there's still room for more optimizations(profiling shows current work has a relatively higher portion of branch mis-predictions, which can affect the performance a lot), but this PR is quite large and let's leave more optimization work to the future.