@quake 2019-05-06
It's awesome but I don't understand assembly code, let's test it with ckb :)